viewAllPosts base modal UI added
