+++ pages/e/components/socialFeed.tsx
@@ -1,15 +1,5 @@
 import { useEffect } from 'react'
-import {
-  startOfToday,
-  differenceInCalendarDays,
-  differenceInCalendarMonths,
-  differenceInYears
-} from 'date-fns'
-import {
-  EventInterface,
-  PostInterface,
-  UserInterface
-} from '../../../shared/interface/common'
+import { EventInterface, UserInterface } from '../../../shared/interface/common'
 import Image from 'next/image'
 import Link from 'next/link'
 import { Button } from '../../../components/buttons/button'
@@ -18,6 +8,7 @@ import useEventValues from '../hooks/event/useEventValues'
 import CreateEventTextInput from './createEventTextInput'
 import { useEvents } from '../../../contexts/events'
 import { EventModalOptions } from '../../../shared/enums/enums'
+import SocialFeedPost from './socialFeedPost'
 
 export default function SocialFeed({
   isMobile,
@@ -107,7 +98,7 @@ export default function SocialFeed({
       <hr />
       {/* use the max-height parameter so it can be resized based from the number of comments. */}
       <div
-        id="social-feed-mobile"
+        id="social-feed-viwer"
         className="thin-scrollbar h-[500px] overflow-y-auto py-[5px]"
       >
         <div className="space-y-[5px]">
@@ -125,109 +116,3 @@ export default function SocialFeed({
   )
 }
 
-/**
- * --it uses the avatar, username, uid for accessing their profile page,
- *   the content of the post. (currently use only text, later check
- * how to add images and gifs in each post and in the text area)
- */
-
-function SocialFeedPost({
-  isMobile,
-  post
-}: {
-  isMobile: boolean
-  post: PostInterface
-}) {
-  const calculateAgeOfPost = (dateTime: Date) => {
-    const differenceDays = differenceInCalendarDays(startOfToday(), dateTime)
-    const differenceMonths = differenceInCalendarMonths(
-      startOfToday(),
-      dateTime
-    )
-    const differenceYears = differenceInYears(startOfToday(), dateTime)
-    if (differenceYears >= 1) {
-      if (differenceYears > 1) {
-        return `${differenceYears}` + ' years ago'
-      }
-      return 'a year ago'
-    } else if (differenceMonths >= 1) {
-      if (differenceMonths > 1) {
-        return `${differenceMonths}` + ' months ago'
-      }
-      return 'a month ago'
-    }
-    if (differenceDays >= 30) {
-      if (differenceDays <= 30 && differenceDays >= 2) {
-        return `${differenceDays}` + ' days ago'
-      }
-      if (differenceDays == 1) {
-        return 'yesterday'
-      }
-    }
-    return 'today'
-  }
-
-  return isMobile ? (
-    <div className="flex flex-col items-start space-x-2  px-[3px]">
-      <div className="flex space-x-2">
-        <Link href={`/u/${post.uid}`}>
-          <div className="relative h-[25px] w-[25px] rounded-full hover:cursor-pointer ">
-            <Image
-              src={post.avatar ?? ''}
-              loading="lazy"
-              layout="fill"
-              objectFit="cover"
-              className="rounded-full bg-gray-200"
-            />
-          </div>
-        </Link>
-        <div className="flex flex-col items-start space-y-0">
-          <div className="my-0 text-[21px]"></div>
-          <p className="my-0 py-0 text-[15px]">
-            <Link href={`/u/${post.uid}`}>
-              <span className="font-bold hover:cursor-pointer hover:underline">
-                {post.username}
-              </span>
-            </Link>
-            commented
-          </p>
-          <div className="my-0 py-0 text-[12px]">
-            {calculateAgeOfPost(post.date_posted)}
-          </div>
-        </div>
-      </div>
-      <div className="">{post.post_content}</div>
-    </div>
-  ) : (
-    <div className="border-1 mr-[5px] flex flex-col rounded-2xl bg-white py-[8px] px-[8px] shadow-sm">
-      <div className="flex items-start space-x-2 ">
-        <div className="relative h-[45px] w-[45px] rounded-full hover:cursor-pointer ">
-          <Link href={`/u/${post.uid}`}>
-            <Image
-              src={post.avatar ?? ''}
-              loading="lazy"
-              layout="fill"
-              objectFit="cover"
-              className="rounded-full bg-gray-200"
-            />
-          </Link>
-        </div>
-        <div className="flex w-fit flex-col items-start space-y-0">
-          <div className="my-0 text-[21px]"></div>
-          <p className="my-0 py-0 text-[15px]">
-            <Link href={`/u/${post.uid}`}>
-              <span className="font-bold hover:cursor-pointer hover:underline">
-                {post.username}
-              </span>
-            </Link>{' '}
-            commented
-          </p>
-          <div className="my-0 py-0 text-[15px]">
-            {calculateAgeOfPost(post.date_posted)}
-          </div>
-          <div className="">{post.post_content}</div>
-        </div>
-      </div>
-    </div>
-  )
-}
